Fix slot_mapping stream ordering and send pool dtype
_ensure_slot_mapping_stream was accessing a non-existent attribute
(storage_backend instead of gpu_connector), causing the fallback to
always fire and enqueue slot_mapping H2D on an orphaned stream with no
ordering relationship to store_stream or load_stream.

Replace with _get_connector_stream(kind) that reads from
lmcache_engine.gpu_connector directly. Load-path callers pass
kind='load' (ordering H2D copy before KV transfer on load_stream);
store-path callers pass kind='store' (ordering before D2H on
store_stream). Fallback uses current_stream, which is always safe.

Also hardcode _send_pool_buf dtype to torch.uint8 to match
_recv_pool_buf (raw_tensor is always uint8 in practice).

+++ lmcache/v1/cache_engine.py
@@ -1741,7 +1741,7 @@ class LMCacheEngine:
                 ):
                     new_size = int(1.5 * need)
                     self._send_pool_buf = torch.empty(
-                        new_size, dtype=raw_tensor.dtype, device=send_device
+                        new_size, dtype=torch.uint8, device=send_device
                     )
                 tensor_to_broadcast = self._send_pool_buf[
                     self._send_pool_cursor : self._send_pool_cursor + chunk_size
